админ часть 24

+++ database/migrations/2021_05_04_174006_create_tasks_table.php
@@ -21,7 +21,7 @@ class CreateTasksTable extends Migration
             $table->integer('order_id');
             $table->string('step');
             $table->timestamps();
-            $table->tinyInteger('user_id'); // кто работает(админ)
+            $table->integer('user_id'); // кто работает(админ)
             $table->tinyInteger('open')->default(1);
         });
     }
